+++ packages/gatsby-admin/src/components/layout.tsx
@@ -1,9 +1,10 @@
 /** @jsx jsx */
 import { jsx } from "theme-ui"
 import { Flex } from "strict-ui"
+import { Helmet } from "react-helmet"
 import Providers from "./providers"
 import Navbar from "./navbar"
-import { Helmet } from "react-helmet"
+import { PageviewTracker } from "./pageview-tracker"
 
 function Layout({ children }: { children: React.ReactNode }): JSX.Element {
   return (
@@ -12,6 +13,7 @@ function Layout({ children }: { children: React.ReactNode }): JSX.Element {
         <title>Gatsby Admin</title>
         <html lang="en" />
       </Helmet>
+      <PageviewTracker />
       <Flex
         gap={0}
         flexDirection="column"